+++ src/admin/dao/index.ts
@@ -1,0 +1,2 @@
+export * from './delete-users.dao';
+export * from './suspended-users.dao';
